Convert CalculatedAdjustments into a proper Concern and remove class_eval
