+++ app/models/spree/payment_method.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 require 'concerns/payment_method_distributors'
-require 'spree/core/calculated_adjustments'
 
 module Spree
   class PaymentMethod < ApplicationRecord
-    include Spree::Core::CalculatedAdjustments
+    include CalculatedAdjustments
     include PaymentMethodDistributors
 
     acts_as_taggable
